Trait cynic::coercions::CoercesTo

pub trait CoercesTo<T> { }

Determines whether a type can be coerced into a given schema type.

Users should not usually need to implement this, it’s handled automatically by the various derives.