Module cynic::schema

Types for describing a GraphQL schema in Rust.

The use_schema macro will mostly output types that make use of the traits in this module to describe the schema. The derives then combine these types with the QueryBuilder type to enforce the restrictions of the schema.

Note that this module is mostly concerned with the marker types output by use_schema, not the actual types users work with. The traits will usually be implemented on marker types, the geneirc parameters will usually be marker types and the associated types will also usually be markers.

Traits

Field

Indicates that a struct represents a Field in a graphql schema.

HasArgument

Indicates that a field has an argument

HasField

Indicates that a type has a given field

HasInputField

Indicates that an input object has a given field

HasSubtype

Indicates that a type has a subtype relationship with another type

InputObjectMarker

Indicates that a type is an InputObject

IsScalar

Indicates that a type is a scalar that maps to the given schema scalar.

MutationRoot

A marker trait that indicates a particular type is at the root of a GraphQL schemas mutation hierarchy.

NamedType

A marker type with a name.

QueryRoot

A marker trait that indicates a particular type is at the root of a GraphQL schemas query hierarchy.

SubscriptionRoot

A marker trait that indicates a particular type is at the root of a GraphQL schemas subscription hierarchy.