Trait cynic::schema::IsScalar

pub trait IsScalar<SchemaType> {
    type SchemaType;
}

Indicates that a type is a scalar that maps to the given schema scalar.

Note that this type is actually implemented on the users types.

Required Associated Types

type SchemaType

The schema marker type this scalar represents.