Trait cynic::QueryFragment

pub trait QueryFragment: Sized {
    type SchemaType;
    type VariablesFields: QueryVariablesFields;

    const TYPE: Option<&'static str> = None;

    // Required method
    fn query(
        builder: SelectionBuilder<'_, Self::SchemaType, Self::VariablesFields>
    );

    // Provided method
    fn name() -> Option<Cow<'static, str>> { ... }
}

A trait that marks a type as part of a GraphQL query.

This will usually be derived, but can be manually implemented if required.

Required Associated Types

type SchemaType

The type in a schema that this QueryFragment represents

type VariablesFields: QueryVariablesFields

The variables that are required to execute this QueryFragment

Provided Associated Constants

const TYPE: Option<&'static str> = None

The name of the type in the GraphQL schema

Required Methods

fn query(builder: SelectionBuilder<'_, Self::SchemaType, Self::VariablesFields>)

Adds this fragment to the query being built by builder

Provided Methods

fn name() -> Option<Cow<'static, str>>

The name of this fragment, useful for operations, maybe fragments if we ever support them…